INCOMPLETE: add support for order-preserving network flows
This commmit addresses feature request
https://github.com/stateright/stateright/issues/4 and should also help
improve the performance for
https://github.com/stateright/stateright/issues/18. Regarding the
latter, using this feature my laptop is able to model check
https://github.com/Nightcro/ble-stateright with 3 nodes, which it could
not do prior.

The data structure to support this is a bit more complex, so I
anticipated additional per-state overhead, which might noticeably impact
check time for models that do not take advantage of the possible state
space reduction by assuming an ordered network for directed flows.
Unfortunately at the moment, the performance difference is larger than
anticipated as the `linearizable-register` example ends up with a larger
state space for some reason. This needs to be resolved before the change
is merged to the `master` branch.

Before:

    == paxos check 6 ==
    Done. states=18972647, unique=9357525, sec=16
    Done. states=18972647, unique=9357525, sec=17
    Done. states=18972647, unique=9357525, sec=18
    == single-copy-register check 4 ==
    Done. states=731789, unique=400233, sec=6
    Done. states=731789, unique=400233, sec=6
    Done. states=731789, unique=400233, sec=6
    == linearizable-register check 2 ==
    Done. states=194646243, unique=41138199, sec=78
    Done. states=194646243, unique=41138199, sec=79
    Done. states=194646243, unique=41138199, sec=79

After:

    == paxos check 6 ==
    Done. states=18972647, unique=9357525, sec=19
    Done. states=18972647, unique=9357525, sec=19
    Done. states=18972647, unique=9357525, sec=19
    == single-copy-register check 4 ==
    Done. states=731789, unique=400233, sec=5
    Done. states=731789, unique=400233, sec=5
    Done. states=731789, unique=400233, sec=5
    == linearizable-register check 2 ==
    Done. states=955073804, unique=209614941, sec=455
    Done. states=955073804, unique=209614941, sec=465
    Done. states=955073804, unique=209614941, sec=469